refactor: move ping logic to repeater server class
